adding shopping cart example
