+++ examples/shopping-cart/streams.js
@@ -1,6 +1,7 @@
 import { mergeSources, mergePlans, stream, plan } from "react-streams"
 import { of, pipe } from "rxjs"
 import { shareReplay, map, tap } from "rxjs/operators"
+import { addToCart, removeFromCart, checkout } from "./plans"
 
 const products = [
   { id: 1, title: "iPad 4 Mini", price: 500.01, inventory: 2 },
@@ -11,69 +12,12 @@ const products = [
 const products$ = of({ products }).pipe(shareReplay(1))
 const cart$ = of({ cart: [] }).pipe(shareReplay(1))
 
-const addToCartPipe = map(id => ({ products, cart }) => {
-  const itemIndex = products.findIndex(({ id: _id }) => _id === id)
-  const cartItemIndex = cart.findIndex(({ id: _id }) => _id === id)
-
-  const item = products[itemIndex]
-  const updateProductItem = { ...item, inventory: item.inventory - 1 }
-
-  const cartItem = cartItemIndex > -1 ? cart[cartItemIndex] : null
-
-  return {
-    products: [
-      ...products.slice(0, itemIndex),
-      updateProductItem,
-      ...products.slice(itemIndex + 1)
-    ],
-    cart: cartItem
-      ? [
-          ...cart.slice(0, cartItemIndex),
-          {
-            ...cartItem,
-            quantity: cartItem.quantity + 1
-          },
-          ...cart.slice(cartItemIndex + 1)
-        ]
-      : [...cart, { ...item, quantity: 1 }]
-  }
-})
-
-const removeFromCartPipe = map(id => ({ products, cart }) => {
-  const itemIndex = products.findIndex(({ id: _id }) => _id === id)
-  const cartItemIndex = cart.findIndex(({ id: _id }) => _id === id)
-
-  const item = products[itemIndex]
-  const updateProductItem = { ...item, inventory: item.inventory + 1 }
-
-  const cartItem = cart[cartItemIndex]
-  const updateCartItem = { ...cartItem, quantity: cartItem.quantity - 1 }
-
-  return {
-    products: [
-      ...products.slice(0, itemIndex),
-      updateProductItem,
-      ...products.slice(itemIndex + 1)
-    ],
-    cart: updateCartItem.quantity
-      ? [
-          ...cart.slice(0, cartItemIndex),
-          updateCartItem,
-          ...cart.slice(cartItemIndex + 1)
-        ]
-      : [...cart.slice(0, cartItemIndex), ...cart.slice(cartItemIndex + 1)]
-  }
-})
-
-const addToCart = plan(addToCartPipe)
-
-const checkout = plan()
-const removeFromCart = plan(removeFromCartPipe)
-
 const store$ = mergePlans(
-  { addToCart, checkout, removeFromCart },
+  { addToCart, removeFromCart, checkout },
   mergeSources(products$, cart$)
 )
+
+export const DebugStream = stream(store$)
 
 export const ProductsStream = stream(store$, map(({ cart, ...rest }) => rest))
 
@@ -84,7 +28,7 @@ export const CartStream = stream(
       .reduce((total, item) => total + item.price, 0)
       .toFixed(2)
 
-    const error = "no error"
+    const error = ""
     const checkoutPending = false
     return { products, total, error, ...rest }
   })
